After statement unit shenanigans
I changed the arrays in unitMethodMapping to have a "D" before all of the distance based units. By not having a "D" at the start of future added time based units, we should be able to differentiate between the two of them. I am not entirely sure what Chloe meant by saying 'Adapt the code to replace the code' so I just tried to do what I could. As far as I can tell, there currently are no functions to get time code so I do not know what would be called in unitMethodMapping?